various clean :
 - take into account @prudhomm review
 - clean timeset.hpp
 - fix interpolate P1 Lagrange optimization with tensor2 symm
