+++ feelpp/feel/feelcore/traits.hpp
@@ -118,5 +118,25 @@ decltype(auto) remove_shared_ptr_f( T&& e )
                      []( auto&& x ) { return x; } )( std::forward<T>(e) );
 
 }
+
+template <class T>
+struct is_std_vector : mpl::false_ {};
+template <class T>
+struct is_std_vector<std::vector<T> > : mpl::true_ {};
+template<class T>
+constexpr bool is_std_vector_v = is_std_vector<T>::value;
+template <class T>
+struct remove_std_vector
+{
+    typedef T type;
+};
+template <class T>
+struct remove_std_vector<std::vector<T> >
+{
+    typedef T type;
+};
+template<typename T>
+using remove_std_vector_t = typename remove_std_vector<T>::type;
+
 } // namespace Feel
 #endif
